[chore/#88] 기존 모델 삭제 및 목 서비스 구현
