+++ Roomie/Roomie/Presentation/Map/Model/MarkerDetailInfo.swift
@@ -1,0 +1,20 @@
+//
+//  MarkerDetailInfo.swift
+//  Roomie
+//
+//  Created by 예삐 on 1/21/25.
+//
+
+import Foundation
+
+struct MarkerDetailInfo {
+    let monthlyRent: String
+    let deposit: String
+    let occupancyType: String
+    let location: String
+    let genderPolicy: String
+    let locationDescription: String
+    let isPinned: Bool
+    let moodTag: String
+    let contractTerm: Int
+}
