+++ Roomie/Roomie/Presentation/Map/Model/MarkerInfo.swift
@@ -1,0 +1,14 @@
+//
+//  MarkerInfo.swift
+//  Roomie
+//
+//  Created by 예삐 on 1/21/25.
+//
+
+import Foundation
+
+struct MarkerInfo {
+    let houseID: Int
+    let x: CGFloat
+    let y: CGFloat
+}
